Get Kalman filter working for 2D canon problem
Still assuming the control vector and sensor vector are the same dimensionality
of state vector. Changing these values will have ramifications for the C vector.
We might be able to get away with keeping all of these vectors the same size, so
I'm leaving it until it gives us problems.